refactor(mapgen-studio): Water proof ledger → Water stats section; P7 ledger entry
The ExplorePanel hydrology surface is now a stats section: one compact line per data family showing semantic counts (plan vs engine, lakes intent vs engine, floodplains applied/rejected) with divergence counts (mismatch/reject/drift) warning-tinted when nonzero, collapsed summary 'matches baseline' / 'N mismatched', and layer-jump chips kept. Claim-status dots/words, lane eyebrows, and acceptance/rendered bookkeeping rows removed — claim semantics stay in the riverLakeInspector module and project docs. Renamed WaterProofSection→WaterStatsSection (and waterProofExpanded→waterStatsExpanded through ExplorePanel/viewStore/StudioShell); system.md amended with the proof-vocabulary ban.

Also appends the P7 deep-semantic-review ledger to docs/projects/mapgen-studio-redesign/00-GOAL.md (slices, live proofs, deferred engine-error follow-up).

Gates: studio tsc clean; bun run test 239/50 green (8 rewritten stats pins).
